+++ Presentation/Controllers/UserController.cs
@@ -1,6 +1,5 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using System.Security.Claims;
 using TailBuddys.Application.Interfaces;
 using TailBuddys.Core.Models;
 using TailBuddys.Core.Models.SubModels;
@@ -33,11 +32,12 @@ namespace TailBuddys.Presentation.Controllers
         }
 
         [HttpPost("login")]
-        // להוסיף לוג.אין מודל
-        // יהיה בנוי מאימייל + סיסמא או לחלופין אימייל + מגוגל איי.די
-        // להוסיף קבלת סיסמא בהתחברות
         public async Task<IActionResult> Login(LoginModel loginModel)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
             string? result = await _userService.Login(loginModel);
             if (result == null)
             {
@@ -50,10 +50,6 @@ namespace TailBuddys.Presentation.Controllers
         [Authorize(Policy = "MustBeAdmin")]
         public async Task<IActionResult> GetAllUsers()
         {
-            Claim? userIdClaims = HttpContext.User.Claims.FirstOrDefault(c => c.Type == "id");
-            card.UserId = userIdClaims.Value;
-
-            // לבדוק ברמת הקונטרולר שהמשתמש תקין ומורשה לבקשה
             List<User> result = await _userService.GetAll();
             if (result == null)
             {
@@ -63,40 +59,64 @@ namespace TailBuddys.Presentation.Controllers
         }
 
         [HttpGet("{id}")]
+        [Authorize]
         public async Task<IActionResult> GetUserById(string id)
         {
-            User? result = await _userService.GetOne(id);
-            if (result == null)
+            string? userId = HttpContext.User.Claims.FirstOrDefault(c => c.Type == "id")?.Value;
+            string? isUserAdmin = HttpContext.User.Claims.FirstOrDefault(c => c.Type == "IsAdmin")?.Value;
+
+            if (isUserAdmin == "True" || userId == id)
             {
-                return BadRequest();
+                User? result = await _userService.GetOne(id);
+                if (result == null)
+                {
+                    return BadRequest();
+                }
+                return Ok(result);
             }
-            return Ok(result);
+            else return Unauthorized();
         }
 
         [HttpPut("{id}")]
+        [Authorize]
         public async Task<IActionResult> Put(string id, [FromBody] User user)
         {
             if (!ModelState.IsValid)
             {
                 return BadRequest(ModelState);
             }
-            User? result = await _userService.Update(id, user);
-            if (result == null)
+            string? userId = HttpContext.User.Claims.FirstOrDefault(c => c.Type == "id")?.Value;
+            string? isUserAdmin = HttpContext.User.Claims.FirstOrDefault(c => c.Type == "IsAdmin")?.Value;
+
+            if (isUserAdmin == "True" || userId == id)
             {
-                return BadRequest();
+                User? result = await _userService.Update(id, user);
+                if (result == null)
+                {
+                    return BadRequest();
+                }
+                return Ok(result);
             }
-            return Ok(result);
+            else return Unauthorized();
         }
 
         [HttpDelete("{id}")]
+        [Authorize]
         public async Task<IActionResult> Delete(string id)
         {
-            User? result = await _userService.Delete(id);
-            if (result == null)
+            string? userId = HttpContext.User.Claims.FirstOrDefault(c => c.Type == "id")?.Value;
+            string? isUserAdmin = HttpContext.User.Claims.FirstOrDefault(c => c.Type == "IsAdmin")?.Value;
+
+            if (isUserAdmin == "True" || userId == id)
             {
-                return BadRequest();
+                User? result = await _userService.Delete(id);
+                if (result == null)
+                {
+                    return BadRequest();
+                }
+                return Ok(result);
             }
-            return Ok(result);
+            else return Unauthorized();
         }
     }
 }
